src/backend/requirements.txt
Flask
Flask-Cors
python-dotenv
supabase
requests
gunicorn